[pycarous] bug fix

+++ Python/pycarous/BatchGSModule.py
@@ -443,8 +443,8 @@ class BatchGSModule():
         trk = np.fmod(angle,360) * 1e2
         gs = np.sqrt(vn**2 + ve**2) * 1e2
         vs = - vd * 1e2
-        self.master.mav.adsb_vehicle_send(0,position[0]*1e7,position[1]*1e7,0,position[2]*1e7,
-          trk,gs,vs,callsign,0,0,0,0)
+        self.master.mav.adsb_vehicle_send(0,int(position[0]*1e7),int(position[1]*1e7),0,int(position[2]*1e7),
+          int(trk),int(gs),int(vs),callsign,0,0,0,0)
 
         """
         self.master.mav.command_long_send(
